# Backend/swagger.yml
openapi: 3.0.0
info: 
  title: Restaurant -API
  version: 1.0.0
  description: API documentation generated by Swagger
paths:
  /menu:
    get:
      summary: Get all orders
      responses:
        200:
          description: A successful response
  /orders:
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required:
                  - name
                  - price
                  - type
                properties:
                  name:
                    type: string
                  price:
                    type: number
                  type:
                    type: string
      responses:
        201:
          description: Order created successfully
        400:
          description: Invalid data supplied
  /orders/{id}:
    delete:
      summary: Delete an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Order deleted successfully
        400:
          description: Invalid ID supplied
  orders/{id}:
    patch:
      summary: Update an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required:
                  - name
                  - price
                  - type
                properties:
                  name:
                    type: string
                  price:
                    type: number
                  type:
                    type: string
      responses:
        200:
          description: Order deleted successfully
        400:
          description: Invalid ID or Data supplied

 